Improvements to delete support of TextBlocks.
The functionality of delete and backspace has been fixed to some degree.
Backspace deletions and normal deletions now work as expected and addtionally,
it is now possible to delete the final character of a line and cause it to
combine with the next line or to backspace while in the first position and
cause that line to combine with the preceding line.