Enable mouse cursor placement on first click in a TextEdit widget.
* A similar improvement that was made to the TextEntry widget was ported over
to the TextEdit widget.
NOTE: While this is an improvement over the former state of this widget, there
are still some remaining issues. For one thing, the MouseEvent handler code
only allows the cursor to be moved to locations within the first line of text
within the widget. This needs to be fixed in the future.