More fixes to the `Object__Free()` functions
There is also a simultaneous effort to update the `Element_CaptureEvent()`
behavior. I had created a program in which the activation of a button destroyed
destroyed a set of elements several layers up (including the button). The
problem is, the event handler system didn't know that the button and the
containers layers up from that  had been freed. So the system must be improved.